LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("07:13", null, null) is "07:14"
PASS stepDown("07:13", null, null) is "07:12"
PASS stepUp("07:13", null, null, 10) is "07:23"
PASS stepDown("07:13", null, null, 11) is "07:02"

PASS successfullyParsed is true

TEST COMPLETE
